Set comment prefix once when first called
Rather than re-determining the comment prefix, set it once when
force_comment is first called and return that same value each time the
function is called.

This fixes a bug where the default value of the variable would change
when re-evaluated in inserters that continually re-evaluate the snippet
structure as inputs change. If an inserter adds the commented lines
to the buffer then the function will return a blank string the next
time it is called, which changes the evaluated snippet and removes the
leading comment.